requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv
scikit-learn
prophet
